requirements.txt

numpy>=1.23
pandas>=1.5
tqdm
scikit-learn>=1.1
scipy
pyyaml
pydantic


transformers>=4.30
sentence-transformers>=2.2
torch>=1.13
accelerate
tokenizers
safetensors

hdbscan
faiss-cpu            
umap-learn


streamlit
flask
gunicorn


python-dotenv
nltk
regex
joblib


matplotlib
wordcloud
plotly


kaggle


typing_extensions
